backend/app/core/security.py
import os
from datetime import datetime, timedelta
from typing import Any, Union

from jose import jwt
from passlib.context import CryptContext
from app.core.config import settings

# Password hashing context
pwd_context = CryptContext(schemes=["bcrypt"], deprecated="auto")

# Developer mode bypass
DEV_MODE = os.getenv("DEV_MODE", "false").lower() == "true"

class DevSuperUser:
    id = 1
    email = "dev@seatsync.com"
    full_name = "Developer Superuser"
    is_active = True
    is_verified = True
    subscription_tier = "admin"
    hashed_password = pwd_context.hash("devpassword")


def create_access_token(
    data: dict, expires_delta: Union[timedelta, None] = None
) -> str:
    """
    Create a JWT access token
    """
    to_encode = data.copy()
    if expires_delta:
        expire = datetime.utcnow() + expires_delta
    else:
        expire = datetime.utcnow() + timedelta(minutes=15)
    to_encode.update({"exp": expire})
    encoded_jwt = jwt.encode(to_encode, settings.JWT_SECRET_KEY, algorithm=settings.ALGORITHM)
    return encoded_jwt

def verify_password(plain_password: str, hashed_password: str) -> bool:
    """
    Verify a password against its hash
    """
    if DEV_MODE:
        return True
    return pwd_context.verify(plain_password, hashed_password)

def get_password_hash(password: str) -> str:
    """
    Hash a password
    """
    return pwd_context.hash(password)

def get_current_user_dev_bypass():
    if DEV_MODE:
        return DevSuperUser()
    return None 